Refactor extract SignInButton to 'gui.auth.sign_in' namespace
